+++ lib/helpful_extensions.dart
@@ -1,5 +1,6 @@
 library helpful_extensions;
 
 export 'src/color_mapping.dart';
+export 'src/color_sort.dart';
 export 'src/list_check.dart';
 export 'src/weekday.dart';